Input Actions
Starting to gain a better understanding of meta tables. Using them to create instance of InputActions and Inputs so that we can use keep the same way of defining input contexts that I have been moving towards, and then using those definitions to instantiate new table instances which we can use to actually listen for and broadcast input behavior.